feat: add get() to index
It's now possible to get a file offset from an index via key.

Please not that it's not guaranteed that a key is really matching as
only a prefix of the key is stored. For example, could an actual key
be "abcb", but only its prefix "ab" is stored. Requesting a key
"abxy" would match this prefix, hence returning the offset, although
it doesn't match the actual key "abcd".